pyproject.toml
[project]
name = "fastarch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.42",
    "starlette-admin>=0.15.1",
    "uvicorn>=0.35.0",
]
